Sped up Chunk loading
I noticed that chunks were taking a very long time to create, on the order of 400ms each, most of the time spent in the "PopulateVoxelMap" method. While investigating I saw that we were recreating the array of face enums for every single block. Meaning we were initializing, populating, and marking it for gc for every single block meaning we were doing it 100s of thousands of times.  So I turned this private variable into a static field, and knocking our chunk creation time to 34ms.